# Optimize radio item renderer to be created 50% faster