<!-- If this is your first pull request for Fyne please read the contributor docs at:
https://github.com/fyne-io/fyne/wiki/Contributing.
Be sure that your work is based off `develop` branch. --> 

### Description:
<!-- A summary of the change included and which issue it addresses.
Please include any relevant motivation and background. -->

Instead of allocating one struct for the renderer and one struct for each canvas object that is part of the renderer, we can embed the types into the renderer and just allocate the larger struct once. This makes for less allocations but we are also theoretically helping the processor cache by keeping the items closer together in memory instead of in different locations. I also made sure to avoid setting a few fields twice. Doing this makes the code not only cleaner but also almost 50% faster (likely due to fewer allocations but using the fields later on might also be faster) for creating the renderer.

```
goos: linux
goarch: amd64
pkg: fyne.io/fyne/v2/widget
cpu: Intel(R) Core(TM) i7-7700 CPU @ 3.60GHz
                      │   old.txt   │               new.txt               │
                      │   sec/op    │   sec/op     vs base                │
RadioCreateRenderer-8   664.0n ± 2%   351.4n ± 2%  -47.09% (p=0.000 n=10)

                      │  old.txt   │              new.txt              │
                      │    B/op    │    B/op     vs base               │
RadioCreateRenderer-8   736.0 ± 0%   704.0 ± 0%  -4.35% (p=0.000 n=10)

                      │  old.txt   │              new.txt               │
                      │ allocs/op  │ allocs/op   vs base                │
RadioCreateRenderer-8   9.000 ± 0%   4.000 ± 0%  -55.56% (p=0.000 n=10)
```

### Checklist:
<!-- Please tick these as appropriate using [x] -->

- [x] Tests included.
- [x] Lint and formatter run with no errors.
- [x] Tests all pass.